Changed name from Python Cluster to Fyrd
Python Cluster is taken several times over, I wanted a more creative
name. Fyrd is the Anglo-Saxon word for 'Army', plus it is just a great
word, so I am going with that.

As this is a large change, I am incrementing the version to beta.2. To
do this I changed the release string to handle semantic versioning
better.

I rebuilt the documentation and altered the configuration to work with
readthedocs.io.