Improved queue waiting
Added explicit support for all slurm statuses, and complete mapping for
all torque statuses to slurm statuses.

Also added a res_time option to the config, which is the time that the
queue will wait for jobs in a preempted, suspended, or stopped state to
switch to a recognized state.

Added more explicit output file checking to the wait and get functions,
an exception will now be raised on Job.get() if output files do not
appear, but they have up to 12 seconds to appear after successful queue
termination