Fixes #45: Make Job class accept function kwargs
Previously, we attempted to determine if the 'args' argument to Job was
a dict or tuple, now we just assume it is a tuple by duck-typing it to
be one.

Instead, now Job accepts a kwargs keyword that is only used when
submitting a function, it is passed as **kwargs to the function call.

This commit is tested with the pyenv tests. Those tests had to be
updated due to a bug in setuptools that resulted in old versions being
installed with `pip install .`. We instead now use `python ./setup.py
develop` for the virtualenv install. We also now explicitly create and
delete virtualenvs on every test.