Documentation overhaul and creation of console script.
All scripts that used to be in bin/ have been moved into a single
entrypoint script called `fyrd` and are managed as subcommands instead
of as separate scripts.

The scripts in bin/ have thus all been converted to shell scripts that
are simple aliases of the relevant fyrd subcommands.

I made a logo of a Saxon shield of the kind used in fyrds and
superimposed a cluster tree.

In addition, I have completely overhauled the documentation and added
the logo. The documentation now builds sensibly as html and pdf, with
the new icon nicely presented in both.

Fixes #37
Fixes #40
Fixes #42

I also realized that there were sphinx outputs included in the repo, so
I deleted all of those and changed the structure of the sphinx folder by
completely recreating the sphinx project and moving the new sources in.

To make script outputs easier to read, I now use the tabulate library,
which I have added as a new dependency.