Change assumption on job missing from queue to success.
Previously, if a job was lost from the queue, failure was assumed and an
Exception was raised. However, in the situation where a long job
precedes a short job, it is very possible for a job to have been removed
from the queue. Therefore, we now assume that absence means success.
This is a flawed assumption, however it will work for now. Ultimately, a
background queue check process will be necessary to ensure all data is
harvested.