Adding back local queue: ongoing
I am adding back the local queue completely rewritten. It now uses a
forking process and an SQLAlchemy sqlite DB to track changes. It also
conforms to the new module style completely.

Integration is still ongoing and the submission functions, plus the
command line script, need to be linked up still.

Further, we still need to add a clean_local_db = time argument to the
config and have the database self clean after a given time.